+++ cmd/gaiad/cmd/root_test.go
@@ -5,8 +5,8 @@ import (
 
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 
-	app "github.com/cosmos/gaia/v5/app"
-	"github.com/cosmos/gaia/v5/cmd/gaiad/cmd"
+	app "github.com/cosmos/gaia/v6/app"
+	"github.com/cosmos/gaia/v6/cmd/gaiad/cmd"
 	"github.com/stretchr/testify/require"
 )