+++ cmd/gaiad/main.go
@@ -6,8 +6,8 @@ import (
 	"github.com/cosmos/cosmos-sdk/server"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 
-	app "github.com/cosmos/gaia/v5/app"
-	"github.com/cosmos/gaia/v5/cmd/gaiad/cmd"
+	app "github.com/cosmos/gaia/v6/app"
+	"github.com/cosmos/gaia/v6/cmd/gaiad/cmd"
 )
 
 func main() {